+++ server/nestadia-core/src/cartridge/mapper_066.rs
@@ -18,7 +18,7 @@ impl Mapper066 {
 
 impl Mapper for Mapper066 {
     fn cpu_map_read(&self, addr: u16) -> CartridgeReadTarget {
-        CartridgeReadTarget::PrgRom((self.prg_bank_selector as u16) * 0x8000 + (addr & 0x7FFF))
+        CartridgeReadTarget::PrgRom((self.prg_bank_selector as usize) * 0x8000 + (addr & 0x7FFF) as usize)
     }
 
     fn cpu_map_write(&mut self, _addr: u16, data: u8) {
@@ -26,11 +26,11 @@ impl Mapper for Mapper066 {
         self.prg_bank_selector = (data & 0x30) >> 4;
     }
 
-    fn ppu_map_read(&self, addr: u16) -> u16 {
-        (self.chr_bank_selector as u16) * 0x2000 + addr
+    fn ppu_map_read(&self, addr: u16) -> usize {
+        (self.chr_bank_selector as usize) * 0x2000 + addr as usize
     }
 
-    fn ppu_map_write(&self, _addr: u16) -> Option<u16> {
+    fn ppu_map_write(&self, _addr: u16) -> Option<usize> {
         None
     }
